internal: use the rnode structure to store the rendering state
This commit moves the rendering state management from the ngl_ctx
structure to the rnode structure.

Previously, the rendering state (graphics state + rendertarget
description) was honored in each relevant nodes (graphicconfig, rtt) by
altering the relevant fields in ngl_ctx before calling
ngli_node_prepare() on the node's child. Then the state was restored to
its previous value at the end of the node's prepare function.

With this commit, the rendering state is now honored by directly setting
the state to the current rnode (which corresponds to the current branch
in the node.gl graph). When a leaf node is reached by the prepare
function, the current rnode structure will hold the final rendering
state that can be used by a graphics pipeline. Using this method:
- removes the need to restore the previous state at the end of the
node's prepare function
- avoids altering two fields in ngl_ctx while ngli_node_prepare()
traverses the graph
- allows the final rendering state to be accessed by rendering nodes
later on (update, draw).

Moreover, the graphicconfig node still keeps a local copy of the parent
state which might be useful to support live-changes later on.