internal: use ngli_gctx_transform_cull_mode()
The cull mode specified by the user needs to be adjusted depending on
the backend coordinate system and rendering. For example, a y-flip is
performed on the geometry with the Vulkan and OpenGL offscreen backends.
In this case, the cull_mode needs to be inverted.

Fixes an OpenGL offscreen rendering regression when culling is enabled
introduced by df18098f01ff9795018b23dfc844225f325db527.