io: add precision_{in,out} parameters
Contrary to uniform precisions, the in/out/varying precisions can
mismatch between the vertex and fragment shader (even in GLSL 100), so
they are exposed in two separated precision fields.

The default precision (highp) matches the one we currently use for
injecting the other vertex attributes and uniforms (with the exception
of samplers).

It might be relevant to tweak the default precision between frag and
vertex (with maybe a lower precision for the fragment).