block: do not switch from SSBO type back to UBO
When a block is shared between 2 pipelines, if one of the pipeline
requires an SSBO type (typically if it needs to write to it in a compute
pass), the block type is changed from UBO to SSBO.

This behaviour was working fine to switch from UBO to SSBO, but it was
unfortunately also switching back to UBO if a later pipeline didn't need
SSBO. A typical scenario would be:

                                    .---------.
                                    |  Group  |
                                    `-.-----.-'
                                    0 |     | 1
                                      v     v
    .---------------.   block  .---------. .--------.
    | ResourceProps |<---------| Compute | | Render |
    | writable=True |          `-----.---' `---.----'
    `---------------'                |  block  |
                                     v         v
                                  .---------------.
                                  | Block         |
                                  | layout=std140 |
                                  `---------------'

The Compute node would make the block type to SSBO (because of the
resource property of writable) and the Render would then restore it back
to UBO (because std140 layout doesn't imply writable, and there is no
resource property implying such a thing).

This commit fixes this by only setting the default UBO type when not set
previously.