internal: rename features.h to feature.h
This may conflict with the system features.h.

An easy way to reproduce the compilation failure is:
  CFLAGS=-I. make nodegl-install

This will cause system headers to include features.h from our tree in
priority over their own (/usr/include/features.h typically). This means
we can not move our headers in a dedicated directory (because it would
imply adding a -I. that would take over the system directory, even with
< > includes).