build: switch libnodegl and ngl-tools to meson
This change is mainly motivated by the need for Microsoft Visual Studio
integration. While it currently works with MinGW for our case, it will
cause integration issues with native C++ libraries such as D3D (because
of the different mangling).

Note that MSVC is still not supported, but this is a first step.

The root Makefile is still present to bootstrap the virtual environment,
but the native code submodules (libnodegl and ngl-tools) are ported to
this new build system.

The existing feature list should be completely ported:
- build on Linux/Mac/Windows/Android/iOS
- debug vs release builds
- generate static and shared libraries (using the default_library
  {static,shared} builtin)
- various control options such as external libraries and log tracing
- detect optional external dependencies (including Apple frameworks)
- exporting a restricted symbols list
- generating a correct pkg-config file according to the configuration
  (typically in shared vs static scenarios), with custom variables
- coverage integration
- reduced symbol visibility to only the public interfaces
- special rules to generate the documentation, the OpenGL wrappers and
  the node specifications
- unit test programs

Additionally, we gain a bunch of useful build features that can't be
exhaustively listed here. But to name a few:
- we now have a smarter behavior and more control with regards to
  external mandatory and optional dependencies
- some operations like generating a pkg-config are now less clumsy
- the builtin capability probing support might be convenient for future
  features
- the portability concerns are now abstracted away for the most part
- out of tree build is now supported
- the detection/config and build are now 2 separated phases, so the
  settings are now memorized (DEBUG, CC), preventing at the same time
  the risk of partial state