rendertarget_gl: do not try to delete the default framebuffer
While this won't be an issue with the OpenGL default framebuffer as
rendertarget_gl.id is set to 0 (glDeleteFramebuffers() silently ignores
0), it is on the EAGL platform as the default framebuffer managed by
glcontext_eagl is a regular framebuffer object.

Also properly sets the invalidate attachments depending on whether we are
dealing with the OpenGL default framebuffer or a framebuffer object.

Forgotten in 0f799469b069a5d6b90724ae6cb96c02bdb1dc41.