log: add NGL_LOG_QUIET
NGL_LOG_QUIET is set to a large enough value so that no future log level
shall be larger. This is also the reason no trailing comma is added at
the end (it must be the last log level).

This log level can be useful in scripting scenario where the output is
parsed and we don't want logging to interfere.