build: do not make custom target calls depend on the install of the lib
Most libnodegl targets do not depend on the compilation and installation
of the library. Typically, nodegl-updatespecs only needs to call a
Python script. Similarly, nodegl-updatedoc and nodegl-updatespecs are
independant from the library itself (they share the source references
and flags). Also, if a target does need the installation, the target
chaining within meson handles it for us.

In addition to making the explicit target calls simpler and faster, it
also actually fixes the generation of the headers in
nodegl-updateglwrappers when you decide to remove them beforehand (to
make sure they are re-generated) because it won't try to compile with
the missing headers anymore.