tests/compute: fix compute_particules workgroups count
Before this patch, we had 128x1x1 workgroups of 4x4x1 compute instances,
so a total of 2048 compute instances for 128 particles. This patch
brings down the number of workgroup to 128/4²=8 (4² being the size of a
work group, 4x4x1) so that we have a matching number of compute
instances and particules: 8x1x1 x 4x4x1 = 128.

This issue was detected while experimenting with MoltenVK in the Vulkan
development branch.