tools/desktop: improve listening socket shutdown code
Only calls close() on Unix systems. closesocket() is its Windows
counterpart which works specifically on sockets.

Also calls the shutdown function unconditionally for all the platforms
as we want to disallow further transmissions on the listening socket.
Even though it has different behaviours on different platforms, calling
it unconditionally doesn't hurt. This will also make this part of the
code re-usable for shutting down client connections.