tools/ipc: use an int instead of a ssize_t to store the send()/recv() return
On Windows, send() and recv() return an int, whereas on Unix systems
they return a ssize_t (which is not supported on Windows).

Using an int to store the return of send() and recv() is fine for our
case as we are only reading or writing packets that cannot exceed the
limits of an int.

This makes the code compatible with MSVC.