pgcraft: use explicit locations for vertex attributes and fragment shader outputs
Explicit locations for vertex attributes and fragment shader outputs are
available when in/out layout qualifiers are available. There is no
downside of using explicit locations and we can potentially reduce the
number of OpenGL calls when using this feature (using
glGetAttribLocation() to probe attribute locations is not needed
anymore).