api: add Path, SmoothPath and AnimatedPath nodes
Path is a child of AnimatedPath. Both are separated because a Path could
be used for other purpose than being animated, for example being drawn.
In this situation, we may want to draw and animate the same Path.

The evaluate/derivate/resolve API that work on Animated* nodes could not
be implemented on the AnimatedPath because of the current architecture.
Currently, an evaluate is supposed to work on nodes not associated with
a node.gl context. This is because this API is useful during scene
construction.

The trick currently used is to initialize the keyframes nodes, which
does not involve any allocation (the anim key frame inits only set the
interpolation functions), which we then use to evaluate the expression.
This works just fine whether the nodes are properly initialized or not.

Doing the same with the path would be more complicated, because it
involves allocations (the internal path API does a bunch of necessary
allocation at init).

For a given AnimatedPath.evaluate() call, we would need to allocate
partially the Path node below, do the evaluation, then free everything,
if and only if the node was not already initialized (associated with a
context). Again, we don't want these allocations to conflict with what
happens at the moment of the proper node initialization (when associated
with a node.gl context).

I attempted that, but a new problem arises because the BufferVec3 used
in SmoothPath.points also needed to be initialized for accessing the
data.  This would have made the code way too fragile, so it was not
implemented.