utils: use a normalized path that works across platforms
The main motivation for this PR is to simplify porting ngl-desktop to windows.

WSL paths are not compatible with Windows paths.

For example:
WSL path: /mnt/c/Program\ Files
Windows path: C:\Program Files

This code creates a path that is compatible with both WSL and windows (as well as other platforms).