ci/coverage,linux: don't use -j or -k when calling make tests to run the unit tests
The -k parameter was used when the tests suite was written in Make. It allowed
all the tests to be ran even if an error occurred (while still failing at the
end). It is not useful anymore and can be removed since we are now relying on
Meson (which does not stop tests execution on error).

The -j parameter can also be removed as Meson (using the ninja backend) runs
the tests in parallel. The option is still useful for building the project as
it allows parallel downloads of our dependencies.