build: move root Makefile logic into a dedicated ngl-env script
GNU/Make requires a *nix like environement to work. On Windows this means we
need an overlay such as WSL (or MSYS2 when not using the MSVC toolchain) to
make it work. WSL introduces a large overhead, and adds complexity for managing
our environment because of the jungling between the Powershell and the WSL
shell. Pulling a whole Linux distribution just for one tool (Make) is
completely overkill. Last but not least, the added branches in the Makefile are
becoming a maintainance nightmare.

The number of common denominators between all the platforms is pretty limited:
1. Make/NMake: Microsoft provides a native implementation of Make called NMake.
   Unfortunately, the feature set overlap just isn't enough to be practical.
2. Meson: having a root meson.build was considered but we might have run into
   some issues for bootstrapping the Python virtual environement (Meson depends
   on Python, so having it is like a snake biting its own tail)
3. Python: we already require Python to be installed on the system on both
   platforms, so we can use a standalone script bootstrapping the virtualenv.
   Python provides a native virtual env API (venv) which actually eases this
   process. This is the selected solution.

Note: a 3-line root Makefile still remains as a handy helper. No option
forwarding is implemented.

Note: Python requirement is bumped to 3.8 because we need it at least
for shlex.join() and dirs_exist_ok parameter in shtuil.copytree().

Windows support was entirely done by Matthieu Bouron.

Signed-off-by: Matthieu Bouron <matthieu.bouron@gmail.com>