build: move Makefile into a standalone configure.py script
GNU/Make requires a *nix like environment to work. On Windows this
means we need an overlay such as WSL (or MSYS2 when not using the MSVC
toolchain) to make it work. WSL introduces a large overhead, and adds
complexity for managing our environment because of the juggling between
the Powershell and the WSL shell. Pulling a whole Linux distribution to
support the build is completely overkill. Last but not least, this
removes the extra branches in the Makefile which are becoming a
maintenance nightmare.

Another positive side effect of this commit is that the build
configuration is now memorised between make calls. We also get rid of
the activate sub process in the (now generated) Makefile by relying on
environment variables.

Note: Python requirement is bumped to 3.8 because we need it at least
for shlex.join.

Large part of the Windows support done by Matthieu Bouron.

Signed-off-by: Matthieu Bouron <matthieu.bouron@gmail.com>