animkeyframe: fix elastic_in formula
The elastic_in and elastic_out formulas are not symmetrical, which means
either elastic_in or elastic_out is wrong.

To go from "in" to "out", we can use 1-f(1-t) (see TRANSFORM_OUT).
Following is the direct application of that transformation on
elastic_in to obtain elastic_out:

    elastic_in(t) = -a×2^(10(t-1)) × sin(((t-1)-s)×2π/p)

    elastic_out'(t) = 1-elastic_in(1-t)
                    = 1-(-a×2^(10((1-t)-1)) × sin((((1-t)-1)-s)×2π/p))
                    = a×2^(-10t) × sin((-t-s)×2π/p) + 1

But currently, we have in the code:

    elastic_out(t) = a×2^(-10t) × sin((t-s)×2π/p) + 1

Difference is -t-s vs t-s between elastic_out' and elastic_out. To do
the transform the other way around (from out to in), we can apply the
same symmetrical function 1-f(1-t):

    elastic_in'(t) = 1-elastic_out(1-t)
                   = 1-(a×2^(-10(1-t)) × sin(((1-t)-s)×2π/p) + 1)
                   = -a×2^(-10×(1-((t-1)+1))) × sin((1-((t-1)+1)-s)×2π/p)
                   = -a×2^(10(t-1)) × sin((-(t-1)-s)×2π/p)

Here again the difference is -t'-s vs t'-s (with t'=t-1).

Now that we have elastic_in(t) symmetrical to elastic_out'(t) and
elastic_in'(t) symmetrical to elastic_out(t), we can display them and
compare which of elastic_in() or elastic_out() is incorrect.

Plotting the 4 curves with the following code can help us:

    import matplotlib.pyplot as plt
    from math import pi, sin, asin

    def _elastic_in_cur(t, a, s, p):
        t -= 1
        return -a * 2**(10.0 * t) * sin((t - s) * (2.0 * pi) / p)

    def _elastic_in_alt(t, a, s, p):
        t -= 1
        return -a * 2**(10.0 * t) * sin((-t - s) * (2.0 * pi) / p)

    def _elastic_out_cur(t, a, s, p):
        return a * 2**(-10.0 * t) * sin((t - s) * (2.0 * pi) / p) + 1.0

    def _elastic_out_alt(t, a, s, p):
        return a * 2**(-10.0 * t) * sin((-t - s) * (2.0 * pi) / p) + 1.0

    def _plot_elastic(n=1000, a=1.7, p=0.25):
        if a < 1.0:
            a = 1.0
            s = p / 4.0
        else:
            s = p / (2.0 * pi) * asin(1.0 / a)

        x_data = [i/n for i in range(n + 1)]
        y_in_cur = [_elastic_in_cur(t, a, s, p) for t in x_data]
        y_out_cur = [_elastic_out_cur(t, a, s, p) for t in x_data]
        y_in_alt = [_elastic_in_alt(t, a, s, p) for t in x_data]
        y_out_alt = [_elastic_out_alt(t, a, s, p) for t in x_data]

        fig, ax = plt.subplots()
        ax.plot(x_data, y_in_cur, label='elastic in (cur)')
        ax.plot(x_data, y_out_cur, label='elastic out (cur)')
        ax.plot(x_data, y_in_alt, label='elastic in (alt)')
        ax.plot(x_data, y_out_alt, label='elastic out (alt)')
        ax.set_xlabel('t')
        ax.set_ylabel('value')
        ax.legend()
        ax.grid(linestyle=':')
        ax.set_title(f'elastic easing {a=} {p=}')

        plt.show()

    if __name__ == '__main__':
        _plot_elastic()

Using this, we quickly realize that elastic_out is much better than our
elastic_in because the initial curve doesn't change direction right
away. We can then assume that elastic_in has simply been badly deduced
from elastic_out. It also makes sense because elastic_out is the most
"natural" effect (and thus probably written first): when releasing an
elastic you get a large wave smoothing over time, while the opposite
effect is not "elastic" semantically speaking.

It is also important to note that this changes only affect the
elastic_in effect for amplitudes > 1. With amplitude < 1, the curves
overlap (which is confirmed by the test references not being changed for
the first elastic_in entries, that is, those without an amplitude set).