utils/hooks: right-strip desktop.send_file() return value
The desktop send_file() hook obtains the destination filename by reading
the stdout of a ngl-ipc execution. Since ngl-ipc prints the filename
followed by a line break, we want to strip it.

This is not done at _ipc_exec() level because the complete raw output
might be useful in some circumstances.

This issue was not spotted earlier because file synchronization only
happens with remote control between 2 desktops.